Refactor code to get active tab
