+++ src/modules/peek/Peek.UI/Native/NativeMethods.cs
@@ -43,9 +43,6 @@ namespace Peek.UI.Native
             DDETopic,
         }
 
-        [DllImport("user32.dll", CharSet = CharSet.Auto, ExactSpelling = true)]
-        internal static extern IntPtr GetForegroundWindow();
-
         [DllImport("Shlwapi.dll", SetLastError = true, CharSet = CharSet.Auto)]
         internal static extern HResult AssocQueryString(AssocF flags, AssocStr str, string pszAssoc, string? pszExtra, [Out] StringBuilder? pszOut, [In][Out] ref uint pcchOut);
