+++ src/modules/peek/Peek.UI/NativeMethods.txt
@@ -9,3 +9,4 @@ BringWindowToTop
 ShowWindow
 GetWindowTextLength
 FindWindowEx
+SID_STopLevelBrowser
